+++ ui/dashboard/package.json
@@ -12,10 +12,12 @@
     "axios": "^0.21.1",
     "bulma": "^0.9.1",
     "core-js": "^3.6.5",
+    "js-cookie": "^2.2.1",
     "mqtt": "^4.2.6",
     "vue": "^3.0.0",
     "vue-router": "^4.0.0-0",
-    "vuex": "^4.0.0-0"
+    "vuex": "^4.0.0-0",
+    "vuex-persistedstate": "^4.0.0-beta.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
