+++ ui/dashboard/src/App.vue
@@ -1,82 +1,22 @@
 <template>
-  <nav class="navbar" role="navigation" aria-label="main navigation">
-    <div class="navbar-brand">
-      <router-link class="navbar-item" to="/">
-        <img src="@/assets/upstage.png" width="80" />
-      </router-link>
-
-      <a
-        role="button"
-        class="navbar-burger"
-        aria-label="menu"
-        aria-expanded="false"
-        data-target="navbarBasicExample"
-      >
-        <span aria-hidden="true"></span>
-        <span aria-hidden="true"></span>
-        <span aria-hidden="true"></span>
-      </a>
-    </div>
-
-    <div id="navbarBasicExample" class="navbar-menu">
-      <div class="navbar-start">
-        <router-link class="navbar-item" to="/">Home</router-link>
-        <router-link class="navbar-item" to="/about">About</router-link>
-
-        <div class="navbar-item has-dropdown is-hoverable">
-          <a class="navbar-link"> More </a>
-          <div class="navbar-dropdown">
-            <a
-              class="navbar-item"
-              href="https://github.com/upstage-org/mobilise/"
-              target="_blank"
-            >
-              GitHub
-            </a>
-            <a
-              class="navbar-item"
-              href="https://github.com/upstage-org/documentation"
-              target="_blank"
-            >
-              Documentation
-            </a>
-          </div>
-        </div>
-      </div>
-
-      <div class="navbar-end">
-        <div class="navbar-item">
-          <div class="buttons">
-            <router-link to="/login" class="button is-primary">
-              <strong>Login</strong>
-            </router-link>
-          </div>
-        </div>
-      </div>
-    </div>
-  </nav>
-  <router-view />
+  <NavBar />
+  <div id="main-content">
+    <router-view />
+  </div>
+  <Footer />
 </template>
 
+<script>
+import NavBar from "@/components/NavBar";
+import Footer from "@/components/Footer";
+export default {
+  components: { NavBar, Footer },
+};
+</script>
+
 <style lang="scss">
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-}
-
-#nav {
-  padding: 30px;
-
-  a {
-    font-weight: bold;
-    color: #2c3e50;
-
-    &.router-link-exact-active {
-      color: #42b983;
-    }
-  }
+#main-content {
+  min-height: calc(100vh - 120px);
+  padding: 10px;
 }
 </style>
