+++ ui/dashboard/src/components/Footer.vue
@@ -1,0 +1,19 @@
+<template>
+  <footer class="footer">
+    <div class="content has-text-centered">
+      {{ new Date().getFullYear() }} — <strong>Stage Dashboard</strong>
+    </div>
+  </footer>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style>
+footer.footer {
+  clear: both;
+  position: relative;
+  padding: 1rem;
+}
+</style>
