+++ ui/dashboard/src/services/baseService.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import store from "@/store/store";
+import store from "@/store/index";
 import router from "@/router";
 // import router from "@/router";
