+++ ui/dashboard/src/store/index.js
@@ -1,12 +1,16 @@
 import { createStore } from 'vuex'
+import createPersistedState from 'vuex-persistedstate'
+import auth from "./modules/auth";
+import user from "./modules/user";
+
+const dataState = createPersistedState({
+  paths: ["auth"],
+});
 
 export default createStore({
-  state: {
-  },
-  mutations: {
-  },
-  actions: {
-  },
   modules: {
-  }
+    auth,
+    user,
+  },
+  plugins: [dataState],
 })
